move helper function and types to new `src/directConnections/` subdir
